🌱 Split BlueprintInfoPanels into per-panel modules
Splits web/src/components/mission-control/BlueprintInfoPanels.tsx (848
lines, carrying an eslint-disable max-lines suppression) into four
focused modules, keeping the original path as a barrel so no importer
changes are needed.

- BlueprintInfoPanelsShared.tsx - STATUS_COLORS, STATUS_LABELS, GaugeRow
- ProjectInfoPanel.tsx          - ProjectInfoPanel
- ClusterInfoPanel.tsx          - ClusterInfoPanel
- DeployModeInfoPanel.tsx       - getDependencyNotes, generateDefaultPhases,
                                  DeployModeInfoPanel
- BlueprintInfoPanels.tsx       - re-exports all eight public symbols

All code is moved verbatim; the only removed line is the
eslint-disable max-lines comment. No behaviour change.

Part of #21500
